Share WeaponType flags
