+++ CSSE/CSWeapon.cpp
@@ -1,19 +1,19 @@
 #include "CSWeapon.h"
 
 namespace se::cs {
-	bool Weapon::getMaterialFlag(WeaponMaterialFlag::Flag flag) const {
+	bool Weapon::getMaterialFlag(TES3::WeaponMaterialFlag::Flag flag) const {
 		return (materialFlags & flag) == flag;
 	}
 
 	bool Weapon::getIsSilver() const {
-		return getMaterialFlag(WeaponMaterialFlag::Silver);
+		return getMaterialFlag(TES3::WeaponMaterialFlag::Silver);
 	}
 
 	bool Weapon::getIgnoresResistance() const {
-		return getMaterialFlag(WeaponMaterialFlag::IgnoresNormalWeaponResistance);
+		return getMaterialFlag(TES3::WeaponMaterialFlag::IgnoresNormalWeaponResistance);
 	}
 
 	bool Weapon::isProjectile() const {
-		return weaponType >= WeaponType::Thrown;
+		return weaponType >= TES3::WeaponType::Thrown;
 	}
 }
