+++ CSSE/CSWeapon.h
@@ -2,42 +2,9 @@
 
 #include "CSPhysicalObject.h"
 
+#include "TES3WeaponFlags.h"
+
 namespace se::cs {
-	namespace WeaponType {
-		typedef unsigned char value_type;
-
-		enum WeaponType : value_type {
-			ShortBlade1H = 0x0,
-			LongBlade1H = 0x1,
-			LongBlade2H = 0x2,
-			Blunt1H = 0x3,
-			Blunt2close = 0x4,
-			Blunt2wide = 0x5,
-			Spear2H = 0x6,
-			Axe1H = 0x7,
-			Axe2H = 0x8,
-			Bow = 0x9,
-			Crossbow = 0xA,
-			Thrown = 0xB,
-			Arrow = 0xC,
-			Bolt = 0xD
-		};
-	}
-
-	namespace WeaponMaterialFlag {
-		typedef unsigned int value_type;
-
-		enum Flag : value_type {
-			IgnoresNormalWeaponResistance = 0x1,
-			Silver = 0x2
-		};
-
-		enum FlagBit {
-			IgnoresNormalWeaponResistanceBit = 0,
-			SilverBit = 1
-		};
-	}
-
 	struct Weapon : PhysicalObject {
 		const char* name; // 0x48
 		Script* script; // 0x4C
@@ -59,7 +26,7 @@ namespace se::cs {
 		unsigned int materialFlags; // 0x74
 		Enchantment* enchantment; // 0x78
 
-		bool getMaterialFlag(WeaponMaterialFlag::Flag flag) const;
+		bool getMaterialFlag(TES3::WeaponMaterialFlag::Flag flag) const;
 		bool getIsSilver() const;
 		bool getIgnoresResistance() const;
